Separate JobRequest into separate directory
This tree can be used for protocol objects (ser/des)
